update fetch top rated movies
