+++ src/templates/swiperCard.hbs
@@ -3,6 +3,6 @@
         <img class="swiper-card__image" src="https://image.tmdb.org/t/p/w500{{poster_path}}"
              data-id="{{id}}" alt="{{title}}"/>
       <span class="movie-card__title">{{title}}</span>
-      <span class="movie-card__vote">{{popularity}}</span>
+      <span class="movie-card__vote">{{vote_count}}</span>
     </div>
 {{/each}}
